Sample DB Access (non-Hibernate)
@zadkayde sample db access using non-Hibernate method. automatically creates table in database TEST (default). Please change settings in Beans.xml accordingly.

You can also change the bean to use the third one if you don't want to use MySQL. Explanation is in file itself. Names of the entities will have "$" in front to indicate the use of this bean.

All files are in package "cmoProject.sample". Add "/dbSample" to the web path.