PMOtoCMO DB integration
Entity classes have to be pure POJO (meaning no additional methods).

DTO is used for API where the entity used for db differs slightly.

An additional class is created to accomodate the conversion methods required between the two classes. The methods are written as static to avoid needing to instantiate the object.